# server/openapi.json
{
  "components": {
    "schemas": {
...
      "BackupDirectoryResponse": {
        "description": "`GET /api/admin/backup-directory`",
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "OkResponse": {
        "description": "The uniform acknowledgement for mutations that return no data.",
        "properties": {
          "ok": {
            "description": "Always `true`; failures come back as an [`ErrorResponse`] with a 4xx/5xx.",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "RemoveFolderBody": {
        "properties": {
          "root_path": {
            "type": "string"
          }
        },
        "required": [
          "root_path"
        ],
        "type": "object"
      },
...
      "RescanBody": {
        "properties": {
          "root_path": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "RescanResult": {
        "properties": {
          "tracks_added": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "tracks_added"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "The server's configured `api_key`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Muorg music server HTTP API, shared by the desktop (Tauri), web and Android clients. Everything except `GET /api/health` and `GET /stream/{id}` requires an `Authorization: Bearer <api_key>` header.\n\n`GET /api/tracks` is paginated and defaults to 500 rows — a bare call returns the first page, not the catalog. Follow `X-Total-Count`.",
    "license": {
      "name": ""
    },
    "title": "Muorg Server API",
    "version": "2.42.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/backup-directory": {
      "get": {
        "operationId": "get_backup_directory",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupDirectoryResponse"
                }
              }
            },
            "description": "Absolute backup directory path"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Where pre-write backups are kept on the server.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/clear-cache": {
      "post": {
        "operationId": "clear_cache",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Cache cleared"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Purge soft-deleted rows immediately.",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/admin/remove-folder": {
      "post": {
        "operationId": "remove_folder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFolderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Root removed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Drop a root and soft-delete the tracks under it.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/rescan": {
      "post": {
        "operationId": "rescan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RescanBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RescanResult"
                }
              }
            },
            "description": "Scan finished"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Rescan one root, or every root when no body is sent.",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}